refactor: getMostRecentValue
Removed 'getMostRecent' method because YAGNI. Using 'getMostRecentValue'
instead. Creating a 'NullFact' class is not trivial, so making it easy
to give a default value for unavailable facts was getting too
complicated.
Since the 'getMostRecent' method was only used to get the value, it was
concluded to use that to get the value directly, making it really easy
to give a consistent default value.